generate and save facts from metrics

+++ src/main/kotlin/pl/poznan/put/oculus/oculusfactsservice/model/GrfIrf.kt
@@ -3,7 +3,7 @@ package pl.poznan.put.oculus.oculusfactsservice.model
 import io.swagger.annotations.ApiModel
 
 @ApiModel(description = "GrfIrf of the fact or rule")
-data class GrfIrf (
-        val grf: Float,
-        val irf: Float
+data class GrfIrf(
+        val grf: Double,
+        val irf: Double
 )
